requirements.txt
pytz==2024.1
Django==4.2.13
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
mysqlclient==2.2.4
drf_spectacular==0.27.2
django_cors_headers==4.4.0